fix: needed float tensor for atomic numbers

+++ ocpmodels/datasets/oe62.py
@@ -69,7 +69,7 @@ class OE62Dataset(DGLDataset):
         graph = DGLGraph(adj_list)
         # get coordinates, typecast to single precision
         graph.ndata["pos"] = torch.from_numpy(data.get("pos")).float()
-        graph.ndata["atomic_numbers"] = torch.LongTensor(data.get("atomic_numbers"))
+        graph.ndata["atomic_numbers"] = torch.FloatTensor(data.get("atomic_numbers"))
         # artificially create substrate "tags" to align with OCP
         graph.ndata["tags"] = torch.ones_like(graph.ndata["atomic_numbers"]) * 2
         # get the interatomic distances as well, typecast to single precision
